Prefer pass-by-value over pass-by-reference for public methods
pass-by-value is faster than pass-by-reference for types that are small
and Copy.  This is technically a breaking change since these methods are
public.  However, most crates should compile fine because the only
method argument that's changing is `&self`, which is provided
implicitly.  The only way to get into trouble would be something like
this:

```rust
let ready = Ready::new;
Ready::is_hup(&ready)
```